+++ UnitTests/PhysicsTestContext.h
@@ -35,6 +35,9 @@ public:
 	Body &				CreateFloor();
 
 	/// Create a body and add it to the world
+	Body &				CreateBody(const BodyCreationSettings &inSettings, EActivation inActivation);
+
+	/// Create a body and add it to the world
 	Body &				CreateBody(const ShapeSettings *inShapeSettings, RVec3Arg inPosition, QuatArg inRotation, EMotionType inMotionType, EMotionQuality inMotionQuality, ObjectLayer inLayer, EActivation inActivation);
 
 	// Create a box and add it to the world